+++ crates/wgso/Cargo.toml
@@ -20,6 +20,7 @@ futures.workspace = true
 itertools.workspace = true
 naga.workspace = true
 strip-ansi-escapes.workspace = true
+walkdir.workspace = true
 
 wgso_parser.workspace = true
